chore : massi

+++ spp_web/database/migrations/2023_07_12_210635_create_requests_table.php
@@ -24,7 +24,7 @@ return new class extends Migration
             $table->foreignIdFor(Period::class)->constrained()->cascadeOnDelete();
             $table->foreignIdFor(Unit::class)->constrained()->cascadeOnDelete();
             $table->string('status')->default('requested');
-            $table->integer('volume');
+            $table->float('volume');
             $table->timestamps();
         });
     }
